Extract PDF import preflight and progress modules
