+++ dev-docs/architecture.md
@@ -68,6 +68,7 @@ js/
   context-card-medical-history-editor.js — Medical History editor, conditions, and family history
   context-card-lifestyle-editors.js — lifestyle context editors, goals, lens, and diet contaminant modal
   pdf-import.js     — PDF/image/text import pipeline, AI analysis, confirm/save merge
+  pdf-import-preflight.js / pdf-import-progress.js — import preflight prompts and progress/status UI
   pdf-import-review.js — import review modal rendering, row filtering/mapping UI state
   import-file-input.js — lazy file-picker import binding and import routing
   import-drop-zone.js — lazy import drop-zone binding shared by page shells
