+++ service-worker.js
@@ -111,6 +111,8 @@ const APP_SHELL = [
   '/js/emf.js',
   '/js/image-utils.js',
   '/js/pdf-import.js',
+  '/js/pdf-import-preflight.js',
+  '/js/pdf-import-progress.js',
   '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/export.js',
