+++ tests/test-correctness-phase2.js
@@ -98,6 +98,8 @@ const pwaAppShellAssets = [
   '/js/import-file-input.js',
   '/js/import-drop-zone.js',
   '/js/pdf-import.js',
+  '/js/pdf-import-preflight.js',
+  '/js/pdf-import-progress.js',
   '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/blob-storage.js',
